minor update
The modal framework has been added and needs to be coded to interrupt the sliceout() function so that you can see what you spun and what its description is.  It also happens to be ugly looking, for now.  Invisible black boxes have been added to make the text appear as if it is scrolling smoothly.