The Alignment and Groupings Update
The arrow finally lines up with the scrolling text.  This alone took many sticky notes and somewhere between 2.5-4 hours. long story short, this was the magic formula: 
-1*((masterwheel.rotation.x/(2*Math.PI)*masterwheel.children.length)-masterwheel.children.length)

Lists for games, rifts, curses, hexes, blessings, etc. have been separated.  Currently only games are shown on the wheel, but soon the other elements will be shuffled into play.

Modal background images have been expanded to include several styles that have a color filter applied to line up with the color of the text and slice of the wheel.  Some images are uglier than others and are prone to be removed in a coming update.