Scalable Vector Graphics Update
Modal text has now been replaced with SVG text that allows for manipulating text color, outline, etc. in a way that CSS could not.  Hooray for cosmetic updates!  Also, forgot to mention that the spin velocity if a hex/curse/etc has been rerolled has now been adjusted to move more carefully with many slices of the wheel remaining, and more quickly with fewer slices.  What this means is that the adjusted speed won't skip over 2 or more slices before landing on an available slice.  (this bug was fixed in the last update, as well as the rift reroll bug, which let you land on multiple rifts before the next actual game)