Socket.io Implementation
contains the server and client files.  to execute, open a command prompt and navigate to the directory within the "wheel" folder wherever you unpack it.  The default port to listen on is 25565, because there's plenty of port forwarding tutorials online for minecraft.  You will need to port forward if playing with others outside of your local network.  Happy spinning!